+++ Changelog.md
@@ -4,6 +4,7 @@ OpenCore Changelog
 - Fixed macrecovery utility on Windows, thx @pzhlkj6612
 
 #### v1.0.7
+- Added System KC loading and cross-KC dependency resolution in OcAppleKernelLib
 - Improved `XhciPortLimit` compatibility with macOS Tahoe, thx @laobamac
 - Updated builtin firmware versions for SMBIOS and the rest
 - Migrated to edk2-stable202511